requirements.txt
bcrypt
fastapi
uvicorn
passlib
python-jose[cryptography]
